serverless_functions/test_image_and_text_classification/requirements.txt
Flask==2.0.1
google-cloud-firestore==2.1.3
werkzeug==2.0.1
google-cloud-firestore
google-cloud-vision
google-cloud-language